refactor: SkeletonContainer 컴포넌트 ts 로 migration
